# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Keenthemes API",
        "description": "Keenthemes products Mock API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Keenthemes server"
        }
    ],
    "paths": {
        "/user/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "description": "Get user by id.",
                "operationId": "a1b628826bab1d988073b662c8b62e90",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User with the provided id."
                    },
                    "404": {
                        "description": "User with the provided id is not found."
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "description": "Update user info by id.",
                "operationId": "1e34be9e4566dcb1fe809416db67b7f7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "User name.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "avatar",
                        "in": "query",
                        "description": "User avatar image path.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User email address.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "position",
                        "in": "query",
                        "description": "User position.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "422": {
                        "description": "Not all required filed are provided!"
                    },
                    "200": {
                        "description": "Users data has been successfully updated!"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "description": "Delete user by id.",
                "operationId": "ba695443dc042c28e6c9a590544b1c26",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User has been successfully deleted!"
                    }
                }
            }
        },
        "/user": {
            "put": {
                "tags": [
                    "Users"
                ],
                "description": "Add new user.",
                "operationId": "47070944c509b343baa9d064ea0f89bc",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "User name.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "avatar",
                        "in": "query",
                        "description": "User avatar image path.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User email address.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "position",
                        "in": "query",
                        "description": "User position.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "422": {
                        "description": "Not all required filed are provided!"
                    },
                    "200": {
                        "description": "User has been successfully added!"
                    }
                }
            }
        },
        "/users/query": {
            "get": {
                "tags": [
                    "Users"
                ],
                "description": "Get user query.",
                "operationId": "bc3c2577f5235c826b755d49ee1818f1",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Pagination current page",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "items_per_page",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search key.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort label.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Sort order asc/desc.",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of the users"
                    }
                }
            }
        },
        "/users/query?filter_two_steps=false": {
            "get": {
                "tags": [
                    "Users"
                ],
                "description": "Get filterd users.",
                "operationId": "f88d4678e43ccc6fe41ea81154753269",
                "responses": {
                    "200": {
                        "description": "List of the users"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Users",
            "description": "API Endpoints of Users"
        }
    ],
    "security": [
        []
    ]
}